refactor(mobile): split trpc-persister-api types + builders + import hooks (mobile-62f6)
Tier B ratchet: drop apps/mobile/src LOC cap from 850 to 500.

Split trpc-persister-api.ts (824 LOC) into:

- trpc-persister-api.ts (41) — orchestrator that composes per-domain
  builders via spread.
- trpc-persister-api.types.ts (292) — TRPCClientSubset, FetchFn, Query/
  Mutation primitives, AVATAR_ENCRYPTION_TIER.
- trpc-persister-api.helpers.ts (42) — sha256Hex, defaultFetch.
- trpc-persister-builders/system-and-buckets.ts (142) — system,
  systemSettings, bucket, customFront, member, friend, field.
- trpc-persister-builders/fronting-and-content.ts (152) — frontingSession,
  frontingComment, note, poll.
- trpc-persister-builders/messaging-and-groups.ts (149) — channel, message,
  boardMessage, group.
- trpc-persister-builders/blob-and-refs.ts (95) — blob, importEntityRef.

The cap drop also surfaced import.hooks.ts (622 LOC) as a violator. Split
it by extracting the read-only progress/summary surface:

- import.hooks.ts (494) — useStartImport, useResumeActiveImport,
  useCancelImport, setup helpers; re-exports the progress hooks for
  existing callers.
- import-progress.hooks.ts (158) — useImportJob, useImportProgress,
  useImportSummary + sumTotals/isTerminalStatus/deriveProgressSnapshot.

Public API of import.hooks.ts unchanged; 1366 mobile tests pass.
