+++ tooling/eslint-config/loc-rules.js
@@ -21,7 +21,7 @@ export const locRules = [
   // Tier B — Lockstep (cap = current max + ~3% buffer)
   { files: ["apps/api/src/lib/**/*.ts"], rules: { "max-lines": ["error", { max: 500 }] } },
   { files: ["apps/api/src/ws/**/*.ts"], rules: { "max-lines": ["error", { max: 500 }] } },
-  { files: ["apps/mobile/src/**/*.{ts,tsx}"], rules: { "max-lines": ["error", { max: 850 }] } },
+  { files: ["apps/mobile/src/**/*.{ts,tsx}"], rules: { "max-lines": ["error", { max: 500 }] } },
   { files: ["packages/sync/src/**/*.ts"], rules: { "max-lines": ["error", { max: 750 }] } },
   { files: ["packages/queue/src/**/*.ts"], rules: { "max-lines": ["error", { max: 500 }] } },
   { files: ["packages/import-core/src/**/*.ts"], rules: { "max-lines": ["error", { max: 500 }] } },
